Add a pipeline.yml
I noticed there is no pipeline.yml in the repo, and instead we have all
the steps defined in the buildkite UI with this gem:

    steps:
      - label: ":pipeline:"
        #command: "buildkite-agent pipeline upload"
        command: echo "we need to add the pipeline in the repo"

I thought I'd be a boyscout and leave the pipeline a little better than I found
it. These steps are copied directly from the UI with no changes. Once this
merges, I'll remove them from the UI and change the initial step to do a proper
upload.