Add hook parameter
Add the ability to run this as either an environment hook or pre-command
hook with the former being the default.
The behaviour is chosen via a new hook parameter.

I tried switching from the cultureamp/aws-assume-role plugin to this
one but found steps using the artifacts plugin were failing. It is
because this plugin was running before the artifacts had a chance to
run, and so the role being assumed was not able to access our artifacts
S3 bucket. The plugin we switched from ran as a pre-command hook and
didn't suffer from this problem.

Move the code for the plugin from hooks/environment to a shared
lib/plugin.bash script.
Update hooks/environment to call the shared code if the new hook
parameter is unset (making it the default) or set to environment.
Create hooks/pre-command to call the shared code if the new hook
parameter is set to pre-command.

Add some tests.

Add details to the README.md file for the new hook and provide an
example.